fix(fetch): correct hostent field widths and AmiSSL tag constants
Two bugs surfaced by the first end-to-end fetch test on real Amiga:

1. hostent layout: h_addrtype and h_length are LONG (32-bit) on AmigaOS
   per sdks/NDK3.2R4/.../netinclude/netdb.h, not 'short' as I had in my
   stub. Reading them as 16-bit meant h->h_length got the wrong bytes,
   producing the misleading "Unsupported address type" on every plain
   HTTP request. Fixed by changing the stub's field types to 'long'.

2. AmiSSL tag IDs: I invented constants like (TAG_USER_BASE + 0x82)
   when the actual values from sdks/AmiSSL-v5.26-SDK/.../amissl/tags.h
   are (TAG_USER + 0x01) for SocketBase, (TAG_USER + 0x0c) for
   UsesOpenSSLStructs, etc. OpenAmiSSLTagList rejected the bogus tags
   silently, leaving ssl_base NULL and every HTTPS request with
   "AmiSSL not available in worker." Fixed by using the correct tag
   values from the SDK header (documented inline).

Also corrected the master library version in worker_open_ssl: use
AMISSLMASTER_MIN_VERSION (5, per the v5.26 SDK) instead of my guessed
3, matching amiga_ssl_lib.c's main-task path.

Version bumped 0.071 → 0.072.

+++ library/vbcc/sharedlib_fetch.c
@@ -50,7 +50,11 @@ typedef struct ssl_method_st SSL_METHOD;
  * hand a per-scope `_sb` pointer so the worker's base flows through.
  * ================================================================ */
 
-/* sockaddr for gethostbyname/connect */
+/* sockaddr / hostent — field layouts MUST match bsdsocket's netdb.h.
+ * On AmigaOS h_addrtype and h_length are LONG (32-bit), not short:
+ *   sdks/NDK3.2R4/SANA+RoadshowTCP-IP/netinclude/netdb.h
+ * Getting this wrong silently returns `h_length != 4` → "Unsupported
+ * address type." Learned that the hard way. */
 #pragma amiga-align
 struct sockaddr_in_stub {
     short sin_family;
@@ -59,10 +63,10 @@ struct sockaddr_in_stub {
     char sin_zero[8];
 };
 struct hostent_stub {
-    char *h_name;
+    char  *h_name;
     char **h_aliases;
-    short h_addrtype;
-    short h_length;
+    long   h_addrtype;
+    long   h_length;
     char **h_addr_list;
 };
 #pragma default-align
